% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{excelCol}
\alias{excelCol}
\title{Return columns numbers corresponding to excel columns headers}
\usage{
excelCol(excelColumns, maxCol = 250)
}
\arguments{
\item{excelColumns}{character vector of excel column names}

\item{maxCol}{max number of columns in the excel sheet}
}
\description{
Return columns numbers corresponding to excel columns headers
}
