% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getSubSample}
\alias{getSubSample}
\title{Return a random stratified subsample of a dataframe}
\usage{
getSubSample(data, group, p = 0.5)
}
\arguments{
\item{data}{data frame}

\item{group}{stratification variable}

\item{p}{proportion of cases to return}
}
\description{
Return a random stratified subsample of a dataframe
}
