% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{lr_cmat}
\alias{lr_cmat}
\title{Extract a confusion matrix from a binary glm}
\usage{
lr_cmat(glm_fit)
}
\arguments{
\item{glm_fit}{an glm object with family - 'binomial'}
}
\description{
Extract a confusion matrix from a binary glm
}
