% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mvsum}
\alias{mvsum}
\title{Get multivariate summary dataframe}
\usage{
mvsum(
  model,
  data,
  showN = T,
  markup = T,
  sanitize = T,
  nicenames = T,
  CIwidth = 0.95
)
}
\arguments{
\item{model}{fitted model object}

\item{data}{dataframe containing data}

\item{showN}{boolean indicating whether sample sizes should be shown}

\item{markup}{boolean indicating if you want latex markup}

\item{sanitize}{boolean indicating if you want to sanitize all strings to not break LaTeX}

\item{nicenames}{boolean indicating if you want to replace . and _ in strings with a space}

\item{CIwidth}{level of significance for computing the confidence intervals, default is 0.95}
}
\description{
Returns a dataframe corresponding to a univariate table
}
\keyword{dataframe}
