% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{niceNum}
\alias{niceNum}
\title{Tidy rounding}
\usage{
niceNum(x, digits = 2)
}
\arguments{
\item{x}{figure to be rounded, can be a vector}

\item{digits}{number of decimal places to display to, defaults to 2}
}
\description{
Tidy rounding
}
