% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{outTable}
\alias{outTable}
\title{The output function for the print methods}
\usage{
outTable(
  tab,
  to_indent = numeric(0),
  to_bold = numeric(0),
  caption = NULL,
  showP = FALSE,
  tblWidth = Inf
)
}
\arguments{
\item{tab}{a table to format}

\item{to_indent}{numeric vector the length of nrow(tab) indicating which rows to indent}

\item{to_bold}{numeric vector the length of nrow(tab) indicating which rows to bold}

\item{caption}{table caption}

\item{showP}{is the p-value column being shown?}

\item{tblWidth}{used only for pander tables output to word indicating where wide tables should be split}
}
\description{
The output function for the print methods
}
