% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_lr_check}
\alias{plot_lr_check}
\title{Outputs a ggplot to test whether numeric predictors are linear in logit}
\usage{
plot_lr_check(model)
}
\arguments{
\item{model}{A glm model object to test}
}
\description{
Outputs a ggplot to test whether numeric predictors are linear in logit
}
