% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_po_check}
\alias{plot_po_check}
\title{This will provide a graph to check the PO assumption, according to
Harrell, RMS, pg 337}
\usage{
plot_po_check(polr.fit)
}
\arguments{
\item{polr.fit}{A polr model object to test}
}
\description{
This will provide a graph to check the PO assumption, according to
Harrell, RMS, pg 337
}
