% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{printConfMatrix}
\alias{printConfMatrix}
\title{Nicely display a confusion matrix from the caret package}
\usage{
printConfMatrix(confMtrx, caption = NULL)
}
\arguments{
\item{confMtrx}{a confusion matrix from the caret package}

\item{caption}{table caption}
}
\description{
Nicely display a confusion matrix from the caret package
}
