% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{printConfStats}
\alias{printConfStats}
\title{Nicely display statistics from a confusion matrix from the caret package}
\usage{
printConfStats(
  confMtrx,
  digits = 2,
  what = c(1, 2, 3, 4, 12),
  caption = NULL,
  tblOnly = FALSE
)
}
\arguments{
\item{confMtrx}{a confusion matrix from the caret package}

\item{digits}{number of digitis to display}

\item{what}{which stats to show, see caret package for details}

\item{caption}{table caption}

\item{tblOnly}{should a dataframe or a formatted object be returned}
}
\description{
Nicely display statistics from a confusion matrix from the caret package
}
