% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{printUnivariateTable}
\alias{printUnivariateTable}
\title{Output several univariate models nicely.
The default output is a kable table for use in pdfs or html, but pander tables can be produced
for Word documents by specifying options('doc_type'='doc') in the setup chunk of the markdown document.}
\usage{
printUnivariateTable(
  response,
  covs,
  data,
  adj_cov = NULL,
  caption = NULL,
  showP = T,
  showN = T,
  tableOnly = FALSE,
  removeInf = T,
  HolmGlobalp = FALSE,
  ...
)
}
\arguments{
\item{response}{string vector with name of response}

\item{covs}{character vector with the names of columns to fit univariate models to}

\item{data}{dataframe containing data}

\item{adj_cov}{character vector containing covariates to adjust for in each (no longer univariate) model}

\item{caption}{table caption}

\item{showP}{boolean indicating if p-values should be shown (may only want to show holm-corrected values)}

\item{showN}{boolean indicating if sample sizes should be shown}

\item{tableOnly}{boolean indicating if unformatted table should be returned}

\item{removeInf}{boolean indicating if infinite estimates should be removed from the table}

\item{HolmGlobalp}{boolean indicting if a Holm-corrected p-value should be presented}

\item{...}{other variables passed to uvsum and the table output functions}
}
\description{
Output several univariate models nicely.
The default output is a kable table for use in pdfs or html, but pander tables can be produced
for Word documents by specifying options('doc_type'='doc') in the setup chunk of the markdown document.
}
