% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rmdBibfile}
\alias{rmdBibfile}
\title{Generate a bibfile for an Rmd document}
\usage{
rmdBibfile(bibfile)
}
\arguments{
\item{bibfile}{a mater bib file contianing all the references in the document (Zotero or Mendeley).}
}
\description{
This function will search through the current Rmd document for citations and
extract these from a central bib file to save as a file-specific bibfile.
It will also search through for citations to R packages and include those,
which can be useful to keep track of which version of a package was used.
to R packages and
}
\keyword{citations}
