% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{survmedian_sum}
\alias{survmedian_sum}
\title{Function to nicely display median survival time by group}
\usage{
survmedian_sum(
  data,
  time,
  status,
  group,
  grpLabel = "Group",
  digits = 1,
  caption = NULL,
  tblOnly = FALSE
)
}
\arguments{
\item{data}{datafrmae}

\item{time}{survival time}

\item{status}{indication of censored or observed data}

\item{group}{the variable to group observations by}

\item{grpLabel}{character value describing the grouping variable}

\item{digits}{the number of digitis to display}

\item{caption}{table caption}

\item{tblOnly}{should a dataframe or a formatted object be returned}
}
\description{
Function to nicely display median survival time by group
}
