% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{survtime_sum}
\alias{survtime_sum}
\title{Function to nicely display survival time summaries at discrete times}
\usage{
survtime_sum(
  data,
  time,
  status,
  group,
  survtimes,
  survtimeunit,
  grpLabel = "Group",
  survtimesLbls = NULL,
  digits = 2,
  caption = NULL,
  tblOnly = FALSE,
  showN = FALSE
)
}
\arguments{
\item{data}{datafrmae}

\item{time}{survival time}

\item{status}{indication of censored or observed data}

\item{group}{the variable to group observations by}

\item{survtimes}{numeric vector specifying the times to calculate survival at}

\item{survtimeunit}{the unit of time that should be indicated in the tables (days, years, etc)}

\item{grpLabel}{character value describing the grouping variable}

\item{survtimesLbls}{if supplied, a vector the same length as survtimes with descriptions (useful for displaying years with data provided in months)}

\item{digits}{the number of digitis to display}

\item{caption}{table caption}

\item{tblOnly}{should a dataframe or a formatted object be returned}

\item{showN}{should the sample sizes be displayed}
}
\description{
Function to nicely display survival time summaries at discrete times
}
