% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{excelCol}
\alias{excelCol}
\title{get a column number (or vector) of Excel columns specified as unquoted letters}
\usage{
excelCol(...)
}
\arguments{
\item{...}{unquoted excel column headers ie excelCol(A,CG,AA)}
}
\description{
get a column number (or vector) of Excel columns specified as unquoted letters
}
