% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{printMultivariableTable}
\alias{printMultivariableTable}
\title{Output a multivariable model nicely}
\usage{
printMultivariableTable(
  model,
  data,
  caption = NULL,
  showN = T,
  tableOnly = FALSE,
  HolmGlobalp = FALSE,
  expnt = NULL
)
}
\arguments{
\item{model}{model fit}

\item{data}{data that model was fit on}

\item{caption}{table caption}

\item{showN}{boolean indicating if sample sizes should be shown}

\item{tableOnly}{boolean indicating if unformatted table should be returned}

\item{HolmGlobalp}{boolean indicting if a Holm-corrected p-value should be presented}

\item{expnt}{defaults to NULL can be set to FALSE for log and logit link models}
}
\description{
Output a multivariable model nicely
}
